requirements.txt
boto3==1.21.10
pandas==1.4.1
psycopg2==2.9.3
pytest==7.0.1
PyYAML==5.3.1
requests==2.27.1
selenium==4.1.2
SQLAlchemy==1.4.31
urllib3==1.26.8